Revert "separated internal and external modules"
This reverts commit 71f9c97403a2ac2621d619969b0c88ebba4204dc.
