+++ src/class/index.js
@@ -1,17 +1,10 @@
-// external point for other modules
-export { default as EventEmitter } from './EventEmitter';
+export { default as Web3Factory } from './web3/Web3Factory';
 export { ProviderFactory } from './provider';
-
-export { default as web3 } from './singleton/web3';
-export { dappBridge, ajv, proxyRequest, http, httpIdentity } from './singleton';
-
-export { default as InpageProvider } from './provider/InpageProvider';
-export { LocalStorage, SettingsStorage } from './storage';
-
 export {
   default as TransactionFactory,
 } from './transaction/TransactionFactory';
 export { default as Transaction } from './transaction/Transaction';
+export { default as EventEmitter } from './EventEmitter';
 export { NotificationError } from './error';
 export { default as Wallet } from './wallet';
 export { default as ENSResolver } from './ens';
