+++ src/class/provider/InpageProvider.js
@@ -1,5 +1,5 @@
 import get from 'lodash/get';
-import { EventEmitter } from '@/class/internal';
+import { EventEmitter } from '@/class';
 import { INPAGE_EVENT, INPAGE_ID_PREFIX } from '@/constants';
 
 export default class InpageProvider {
