+++ src/class/singleton/proxyRequest/ProxyRequest.js
@@ -1,5 +1,5 @@
 import { Decorator, ProviderUrlDecorator } from './decorator';
-import { NotificationError } from '@/class/internal';
+import { NotificationError } from '@/class';
 import { IDENTITY_MODE } from '@/constants';
 import { createProvider } from './provider';
