+++ src/class/transaction/TransactionFactory.js
@@ -1,5 +1,4 @@
 import web3 from '@/class/singleton/web3';
-import { Transaction } from '@/class/internal';
 
 import Transaction from './Transaction';
