+++ src/class/wallet/proxy/HDProxy.js
@@ -1,5 +1,5 @@
 import HDKey from 'ethereumjs-wallet/hdkey';
-import { NotificationError } from '@/class/internal';
+import { NotificationError } from '@/class';
 import getChildrenAddress from './utils/getChildrenAddress';
 
 export default class HDProxy {
