+++ src/class/wallet/proxy/hardware/TrezorProxy.js
@@ -1,9 +1,9 @@
 import TrezorConnect from 'trezor-connect';
 import Tx from 'ethereumjs-tx';
 import HDKey from 'ethereumjs-wallet/hdkey';
-import { NotificationError } from '@/class/internal';
-import web3 from '@/class/singleton/web3';
+import { NotificationError } from '@/class';
 import { HARDWARE_DERIVIATION_PATH } from '@/constants';
+import web3 from '@/class/singleton/web3';
 import getChildrenAddress from '../utils/getChildrenAddress';
 
 const { sha3, toHex, toDecimal } = web3.utils;
