+++ src/services/cryptoData.js
@@ -1,6 +1,7 @@
 import throttledQueue from 'throttled-queue';
 
-import { NotificationError, http } from '@/class';
+import { http } from '@/class/singleton';
+import { NotificationError } from '@/class';
 import { cryptoDataValidator } from '@/schema';
 
 const throttle = throttledQueue(3, ENV.serviceThrottleTimeout);
