+++ src/store/accounts/actions.js
@@ -1,5 +1,6 @@
 import { get, mapKeys, isEmpty } from 'lodash';
 import { userService, localSettingsService } from '@/services';
+import web3 from '@/class/singleton/web3';
 import Bip39 from 'bip39';
 import HDKey from 'ethereumjs-wallet/hdkey';
 import EthWallet from 'ethereumjs-wallet';
@@ -9,7 +10,6 @@ import {
   TrezorProxy,
   LedgerProxy,
   HDProxy,
-  web3,
 } from '@/class';
 import { keystore } from '@endpass/utils';
 import { WALLET_TYPE } from '@/constants';
