+++ src/store/dapp/actions.js
@@ -1,5 +1,8 @@
 import Web3Dapp from 'web3-dapp';
-import { NotificationError, InpageProvider, web3, dappBridge } from '@/class';
+import web3 from '@/class/singleton/web3';
+import { NotificationError } from '@/class';
+import { dappBridge } from '@/class/singleton';
+import InpageProvider from '@/class/provider/InpageProvider';
 import { DAPP_WHITELISTED_METHODS } from '@/constants';
 import {
   ADD_REQUEST,
