+++ src/store/plugins/web3.js
@@ -1,6 +1,7 @@
 import Web3 from 'web3';
 
-import { ProviderFactory, web3 } from '@/class';
+import web3 from '@/class/singleton/web3';
+import { ProviderFactory } from '@/class';
 
 const setProvider = ({ dispatch }, provider) => {
   web3.setProvider(provider);
