+++ test/unit/specs/class/singleton/httpIdentity.spec.js
@@ -1,4 +1,4 @@
-import { httpIdentity } from '@/class';
+import { httpIdentity } from '@/class/singleton';
 import store from '@/store';
 import { REQUEST_TIMEOUT_MSEC } from '@/constants';
