+++ test/unit/specs/services/identityMode.spec.js
@@ -1,6 +1,7 @@
 import axios from 'axios';
 import MockAdapter from 'axios-mock-adapter';
-import { NotificationError, httpIdentity, proxyRequest } from '@/class';
+import { NotificationError } from '@/class';
+import { httpIdentity, proxyRequest } from '@/class/singleton';
 import { IDENTITY_MODE } from '@/constants';
 import { addresses } from 'fixtures/accounts';
