+++ test/unit/specs/services/localSettings.spec.js
@@ -1,4 +1,4 @@
-import { LocalStorage } from '@/class';
+import LocalStorage from '@/class/storage/LocalStorage';
 import { STORAGE_USER_META_KEY } from '@/constants';
 
 const localSettings = require.requireActual('@/services/localSettings').default;
