+++ test/unit/specs/services/user.spec.js
@@ -1,6 +1,7 @@
 import MockAdapter from 'axios-mock-adapter';
 
-import { NotificationError, httpIdentity } from '@/class';
+import { NotificationError } from '@/class';
+import { httpIdentity } from '@/class/singleton';
 import { WALLET_TYPE } from '@/constants';
 import { successResponse } from 'fixtures/identity';
