+++ test/unit/specs/store/dapp/actions.spec.js
@@ -5,7 +5,8 @@ import {
   CHANGE_INJECT_STATUS,
 } from '@/store/dapp/mutations-types';
 import { MAIN_NET_ID } from '@/constants';
-import { web3, dappBridge, NotificationError } from '@/class';
+import web3 from '@/class/singleton/web3';
+import dappBridge from '@/class/singleton/dappBridge';
 import { checksumAddress, v3password } from 'fixtures/accounts';
 import {
   commonRequest,
